Reorganize update submodules
