+++ client/verta/verta/deployment/update/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+from ._strategies import (
+    DirectUpdateStrategy,
+    CanaryUpdateStrategy,
+)
